requirements.txt
python-telegram-bot==21.5
anthropic==0.34.0
httpx==0.27.0
